Added QuickMatch option to main menu
* Updated the EOS OnlineSubsystem class and the GameInstance blueprint
to allow creating instances of public matches. A player can search for
a public match to join by pressing the QuickMatch button on the main menu.
The public match requires a new Destination in the Meta Platform services.

* Fixed a bug where the client mesh socket locations were not the same as
the servers. This required to switch on "FPSDebugMode" on the VRCharacter settings
from the VRExpansion Plugin.

* Added ambient soundtracks on the maps.

* Various other small changes...